Adding server inside start
